backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
pandas>=2.2
numpy>=1.26
scipy>=1.12
openai>=1.30
httpx>=0.27
